+++ apps/backend/src/lib/events.tsx
@@ -125,7 +125,7 @@ export async function logEvent<T extends EventType[]>(
       data = await eventType.dataSchema.validate(data, { strict: true, stripUnknown: false });
     } catch (error) {
       if (error instanceof yup.ValidationError) {
-        throw new StackAssertionError(`Invalid event data for event type: ${eventType.id}`, { eventType, data, error, originalData, originalEventTypes: eventTypes }, { cause: error });
+        throw new StackAssertionError(`Invalid event data for event type: ${eventType.id}`, { eventType, data, error, originalData, originalEventTypes: eventTypes, cause: error });
       }
       throw error;
     }