+++ examples/e-commerce/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/e-commerce-demo
 
+## 2.6.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.36
+
 ## 2.6.35
 
 ### Patch Changes