+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/stack-shared
 
+## 2.6.36
+
+### Patch Changes
+
+- Various updates
+  - @stackframe/stack-sc@2.6.36
+
 ## 2.6.35
 
 ### Patch Changes